rename single_dataset_scalar_performance -> single_dataset_performance rename single_dataset_scalar_performance_test ->  single_dataset_performance_test
